Move host command structure to protocol layer
