+++ examples/htool_jtag.c
@@ -52,7 +52,7 @@ static int jtag_read_idcode(struct libhoth_device *dev,
 
   struct ec_response_jtag_read_idcode_operation response;
   size_t response_length = 0;
-  int ret = htool_exec_hostcmd(
+  int ret = hostcmd_exec(
       dev, EC_CMD_BOARD_SPECIFIC_BASE + EC_PRV_CMD_HOTH_JTAG_OPERATION,
       /*version=*/0, &request, sizeof(request), &response, sizeof(response),
       &response_length);
@@ -158,7 +158,7 @@ static int jtag_test_bypass(struct libhoth_device *dev,
 
   struct ec_response_jtag_test_bypass_operation response;
   size_t response_len = 0;
-  ret = htool_exec_hostcmd(
+  ret = hostcmd_exec(
       dev, EC_CMD_BOARD_SPECIFIC_BASE + EC_PRV_CMD_HOTH_JTAG_OPERATION,
       /*version=*/0, &request, sizeof(request), &response, sizeof(response),
       &response_len);
