+++ protocol/BUILD
@@ -1,0 +1,10 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "host_cmd",
+    hdrs = ["host_cmd.h"],
+    srcs = ["host_cmd.c"],
+    deps = [
+        "//transports:libhoth_device",
+    ],
+)
